[Pxr] Update github repository documentation
- Move Issues document into an Issues template. This allows github to
prefill issue requests for users, so they will include all of the information
we want, rather than having it in a separate document they may ignore.
- Create a pull request template in the same vein.
- Add a contributing document; this makes github display a banner
before users can submit a pull request. This is in an effort to
streamline the process of our CLAs. For further information on these see:

https://help.github.com/articles/creating-a-pull-request-template-for-your-repository/
https://help.github.com/articles/creating-an-issue-template-for-your-repository/
https://help.github.com/articles/setting-guidelines-for-repository-contributors/

(Internal change: 1640807)