Fix bug where targets that were authored inside variants and that pointed to relocated prims caused composition errors.
This was due to how the mapping functions between variant
nodes and their parents were set up. Since variants are
conceptually just switches for local opinions, the namespace
for a variant node should be exactly the same as its parent
node. This is represented by an identity mapping function.
However, in this case Pcp was erroneously adding in
relocations to this mapping function. This caused target
paths that pointed to relocated prims to be unmappable,
because the combination of the identity and relocation
mappings would break the bijection between parent and
variant namespaces that PcpMapFunction maintains.

This change makes it so that Pcp no longer adds relocations
to the map functions used by variants. This behavior was
pushed into _CreateMapExpressionForArc, which is called
in exactly the places where relocates are necessary.

(Internal change: 1641091)