Fix dirty instances not being updated.
When we modify an instance we set the dirty bit on the instancer in the change tracker.  However, we go to sync, we start by looking at the rprims.  If the rprim is dirty, then we would look to see if the instancer is dirty and update the instance buffer.

So we need a way to dirty the rprims associated with an instancer when the instancer is dirty.

I've solved this by adding a reverse map to the change tracker, so that we can dirty the rprims.

Changes:
- The change tracker now keeps a set of rprims associated with each instancer.
- There are new entry points in the change tracker to add and remove these associations.
- When an instancer is dirtied, we also set the DirtyInstancer bit on all rprims associated with that instancer.
- The render index is responsible for updating the association with the change tracker when rprims are added and removed.

(Internal change: 1642257)