Fix tripped verify when trying to use a surfaceShader with an error.
Before this fix, when a user passed in a shader to Hydra with an error
in it, we raised a failed verify.  With this change, we now issue a
warning to the user and fallback to the default surface shader.  This
allows the system to continue functioning so that users can more easily
iterate on their hardware shaders.

(Internal change: 1644990)