Getting Python traceback and exception string no longer clobber exception state.
Previously TfPyExceptionState::GetExceptionString() and
TfPyGetTraceback() could clobber the current exception state.
This means raising an error like a coding error that should
bubble up to Python as an exception could instead cause the
exception state to be cleared, either leading to no error in
Python or a SystemError when no result was produced but no
error was set.

Now we have the RAII TfPyExceptionStateScope type that saves and
restores the exception state in its c'tor and d'tor, respectively.
We put one of those around each of the GetExceptionString() and
TfPyGetTraceback() implementations to ensure they don't mess up
the exception state.

(Internal change: 1651753)