Adds shaderAdapter which is an interface that provides shading information.
The main motivation of this code is to be able to provide a shader
network without having to explicitly have hydra look prims.

The previous logic that was in the UsdImagingDelegate is now moved into
the UsdImagingDefaultShaderAdapater.

New shaderAdapters can be registered for particular paths using
AddShaderAdapater().

The function call that normally returns the shader binding in
primAdapter was made virtual.  This allows a primAdapter to override it
to provide a new shaderId for which it can register a shader adapter.

(Internal change: 1652894)