[Hd, Hdx] move camera and light from Hd to Hdx
motivation:
HdLight is a specific implementation of GlfSimpleLight, which we
wanted to move out of hd. Same thing for HdCamera.

fix:
We introduce new type for the base class of non-renderable prims,
HdSprim (state prim). The decendants of this class will be used for
communicating scene delegate and cache data with change tracking.
They don't provide drawItems, thus not to be displayed.
HdxCamera and HdxLight inherit from HdSprim and do the same
job as before.

other changes:
- move DirtyBits for camera and light to Hdx as well.
- move Tokens for camera and light to Hdx as well.
- update DebugCodes accordingly.
- testHdCameraAndLight is moved to testHdxCameraAndLight.
- hd no longer depends on cameraUtil
- add API guards for plugin compatibility.

todo:
- HdxSimpleLightTask and HdxShadowTask have dynamic_cast
  to check if a sprim is HdxLight or not. this doesn't scale well.
  We continue to discuss to improve this (by VtValue, schema or type
  registration etc).

(Internal change: 1653086)