[Usd] Fix for attribute blocking. Previously, the typed Get() API in Usd was broken with respect to blocks. This occured because of failed calls to StoreValue() during the composition phase.
To alleviate this, this adds a member, isValueBlock to the underlying
SdfAbstractData classes, which allows consumers, such as the composition
functions, to query whether or not a fetched value was a block.

This also fixes a bug in the UsdAttributeQuery code, which wasn't consulting
the typed value's "blockedness" at all.

(Internal change: 1653268)