Tf: allow TF_SUPPORTS_WEAKPTR to be applied to final classes.
The previous implementation of __GetTfWeakBase__ member detection
(Tf_HasGetWeakBase) relied on being able to derive from T.  However, this
is not allowed for final classes, resulting in a compilation error when
putting instances of a final class into a TfAny.

In addition to supporting final classes, there are two other changes in behavior:

1. Classes with a private or protected __GetTfWeakBase__ are no longer
considered to support TfWeakPtr.
2. Classes with a member function named __GetTfWeakBase__ that does not
match the required signature are no longer considered to support TfWeakPtr.

The tests have been moved to the TfWeakPtr unit test so they will be run whenever tests
are built.  There is no runtime component; the unit test will simply fail to build.

(Internal change: 1653331)