Pcp: Remove redundant tracking of "used asset paths".
This information is already available and (more richly
represented) in the prim index and error structures.
Furthermore, there is some unspecified nuance involved
in what it means to be "used", so move that policy
downstream to the one (internal) client.

Removing this saves a bit of storage: a pointer in
every PcpPrimIndex plus any allocated std::vectors.

(Internal change: 1653580)