[Hd] packed normals
This change adds support of packed format for vertex normals.
GL_INT_2_10_10_10_REV will be used as a vertex normal format
by default for the coarse (unrefined and unquadrangulated)
vertex normals. It reduces 20~30% of GPU memory consumption
in typical cases.

One caveat though: if we use uniform subdivision (which is default now),
we refine the coarse smooth normals to the dense mesh as an
approximation. This refinement is done by OpenSubdiv as a primvar
refinement, which only supports single-precision float currently.
We use float as before if a mesh needs to be refined, however, if the
mesh has already been drawn with packed normals for the coarse
level, additional 32bit float normals will be populated instead of sharing
same buffers between coarse and dense mesh. It results in consuming
more memory than not using packed type at all. This happens
when a prim is initialized with packed and then is requested to draw
as refined. If the prim is initialized with 32bit float normals from the
beginning, packed normals will never be used for that prim (same
footprint as before).
We'd like to avoid this problem by enabling adaptive subdivision.
(it's desirable for memory anyway)

adding HD_ENABLE_PACKED_NORMALS (defaults to 1) for better
future bug triage.

(Internal change: 1653962)