During NURBS conversion, if normalizeNurbs is turned on then we can get a rounding error when the knot vectors are scaled (for example, an order-2 knot vector can look like (-0.99999999999 0 0.99999999999 1.999999999 despite the double precision math).
However, the knot ranges are always forced to [0,1] when normalized, so the result is that you can get a uRange or vRange that exceeds the actual knots by some extremely small amount.

This results in validation errors downstream in Presto, and currently causes some sort of presto memory corruption (bugs 136704 and 136705) although it seems OK in other parts of the pipeline.

This fix is to actually change the boundary knot values so that they do not fall inside the [0,1] normalized range.

(Internal change: 1656282)