[Hd] prevent shader compilation error when sampling coordinate is not defined
Currently hydra's codeGen generates texture accessor as

HdGet_diffuseColor(vec2 coord) {
   return texture(sampler2d_diffuseColor, coord);
}
HdGet_diffuseColor() {
   return HdGet_diffuseColor(HdGet_uv());
}

but it fails to compile if the prim doesn't have HdGet_uv defined in primvars.

For easier integration, this change adds local fixes
vec2(0) instead if not.
Ideally this should be detected in codegen if the inPrimVar exist or not,
instead of GLSL compile time switch.

(Internal change: 1656980)