Fix for not resetting varying bit after initial frame draw.
The bug report was complaining about poor performance playing back a shot, even though most of the shot was unloaded or invis'ed.

Profiles revealed that the time was being spent in Hydra during Syncing even though only a small number of items were changing in the frame.

The problem was caused by invalidating the prims during the initial renders causing the prims to become marked as varying.

We have a heuristic that resets the varying flag on prims when more than a certain percentage of prims in the dirty list are actually clean.  This percentage was set very high (90%) when the shot it question has 100000's of items, but was only reaching 88%.

Changes:
- Changed the heuristic to be 25%
- Added a debug flag that reports the current heuristics

(Internal change: 1657082)