add support for exporting user-tagged Maya attributes as primvars and usdRi attributes
This change fleshes out the metadata schema for the JSON blob stored in
the 'USD_UserExportedAttributesJson' attribute on Maya nodes to allow
specifying which type of attribute a particular Maya node should be exported
as. Primvars can also have their interpolation specified. It includes
changes to the Maya widget to allow authoring these opinions when tagging
attributes for export.

(Internal change: 1658018)