Avoid destroying MallogTag thread local data.
MallocTag keeps some thread local data around to track the stack
of tags among other things.  We allocated this using "static
thread_local _ThreadData data".  This meant the object was
destroyed when the thread exited.  But controlling the order of
object destruction is hard and some objects destroyed later were
doing memory allocation and deallocation.  That would re-enter
MallocTag and try to access the destroyed _ThreadData.

This change allocates and constructs the _ThreadData in such a
way that we neither use the heap nor destroy the object.  We
unfortunately leak the blocks allocated on the heap by the
members in _ThreadData.

(Internal change: 1658244)