If a shader did not compile we would replace it with a fallback shader however that would not be taken into account when binding buffers and other parts of the pipeline.
To solve this issue now we keeping the shaders used in the program so we can access them anywhere in the batch, this also consolidates the different ways to access the shaders that we had previously.

[Hd] In renderPassState : Function GetShaders() Now it only returns the lighting and renderpass shader, it does not know about surface shader and it does not take into account the override shader, now it is up to the consumer of that function.
In drawBatch : Currently if a shader did not compile you would replace the shader with a fallback shader however that would not be taken into account when binding buffers and other parts of the pipeline. Now we store the current shader we are using and we use that one to bind buffers correctly.
In inmediateDrawBatch and indirectDrawBatch : Both are using the shaders stored in the program associated to a batch.

(Internal change: 1661165)