add a Reset() method to UsdMayaGLBatchRenderer and make sure it's invoked on Maya scene updates
The batch renderer is used as a static singleton object, and it keeps caches
that contain ShapeRenderer objects, each of which may hold a pointer to a
UsdImagingDelegate object. Those delegates maintain a strong pointer to the
UsdStage. As a result, we need to make sure that the batch renderer's state
is cleared so that the stage and any loaded layers can be released.

This resolves an issue where making edits directly in USD with a stage
loaded with an assembly or proxy node resulted in those edits persisting
into a new scene, since the strong pointers in the delegate(s) were
preventing the stage and layers from being released.

(Internal change: 1661166)