Fix for NI-PI USD invalidation.
When USD NI instances PI instances, Hydra does not know about the Rprim associated with the PI instancer.  Therefore, the rprims were not getting marked as dirty when for example a transform occured in the transform heirarchy.

This problem was already being solved for PI cases, by asking the adapter for the dependent set of Rprim's and dirtying those.  However, the adapter interface wasn't implemented for NI instances.

In addition to implementing the required interface, NI instances use a different path to identify the instancer than PI.

Changes:
- Implemented GetDependPaths() for UsdImagingInstanceAdapter.
- In _MarkSubtreeDirty() now get the depend paths for the NI instancer and mark those subtrees as dirty.

(Internal change: 1662332)