Pcp: Add another culling pass to Pcp graph optimization that can be more aggressive.
Node-culling is a crucial optimization pass in Pcp composition that
is a big win against the tendency for scene structure to pile
up down namespace.  Prime contributors are global inherits
(which rarely have opinions below the root level) and sets (which
add another reference at each level of grouping).

The rules for when we can cull nodes are subtle due to the need
to preserve structure required for correct composition, but
it turns out we can be more aggressive once we are finalizing
the graph, vs. using it an intermediate result.

(Internal change: 1662855)