Pcp: Variant node culling attempt #2.
Node-culling is a crucial optimization in Pcp composition that
is a win against the natural tendency for scene structure to
accumulate down namespace.

The basic idea is that, as we walk down namespace, we can prune
off branches of the composition structure as soon as we know
they contribute no further opinions.

However, the rules for which nodes we can cull are subtle due
to certain features that require preserving empty structure.
But it turns out we can be more aggressive with variant culling
than we had been.

(Internal change: 1664515)