Fix for crash in copy computation
The Hydra mesh Rprim sets up a copy computation to combine vertex buffers that have different formats.  When were running the computation, we discovered that the source of the copy didn't have a GPU resource allocated to it.  The reason why their was no GPU resource is because the points array was empty, so because of that the packed smooth normals was producing no output and the memory manager was determing the size was 0, so didn't provide a resource for it.

This all happens in a deferred way, so we don't know at the time of adding the computation that the result is going to be empty, therefore the copy computation is seeing these empty unallocated buffers at execution time.

Changes:
- Reordered the copy computation execution function, so that the empty check comes before the is resource allocated check.

(Internal change: 1664714)