Removing glfq from pxr
glfq is the last Qt C++ dependency in pxr. In pxr, it's being used for enabling debug context
in usdview. it's useful, but we think it doesn't worth to ask Qt dependency just for it,
and we have another way to debug GL errors such as API trace and GL debugger.

For GL debugging in usdview, we'll revisit to add that functionality in another way later on.

(Internal change: 1665204)