cmake: fix DESTDIR support for generated files
We were emitting cmake code that was attempting to write directly into
the installation prefix.

This breaks when building using DESTDIR because we are writing directly
into the prefix without taking the DESTDIR into account.  This breaks
this common use case:

	make DESTDIR=/usr install

Fix the build by writing the generated files into the cmake binary
directory and then using a normal install(...) command to install them.
This lets cmake handle the pathing, and thus DESTDIR will be correctly
handled.

Closes #84
Signed-off-by: David Aguilar <davvid@gmail.com>