Partial Hydra Support for Reloading Shader Programs.
This change provides Hydra support for reloading shader programs for Surface and Geometry shaders.  It doesn't currently support Lighting, Render Pass and Culling shaders.  The engine marks all prims as dirty to trigger an re-evaluation of the shader state, marks all surface shaders as dirty to cause those prims to be synced.  It also removes all the geometric shader instances from the GeometricShaderRegistry.  Finally it causes the Fallback surface shader to be reloaded.
Reloading Renderpass shader is now partially supported as we will correctly compile and re-link the shader if the contents of the file changes, but I'm not currently reloading the glslfxFile.

Changes:
- Added ReloadAllShaders that dirties state so it gets reevaluated and removes cache entries from the resource registry.
- Added ability to remove all entries from the HdInstanceRegistry.
- The Render Index can now reload the fallback surface shader, but creating an new one and swapping it with the old one.
- Renderpass shader now uses the hash of the contents of the file rather than a hash of the filename.

(Internal change: 1665781)