Fix GetLoadSet when loaded prim no longer exists
UsdStage::GetLoadSet asserted that all prims in its load
set will always exist on the stage. This was incorrect; it's
actually very easy to get into such a state -- for example,
by deactivating the ancestor of a loaded prim.

This change removes the assertion and will simply return
the previously-loaded path in this case. This accurately
reflects what payloads are loaded in the underlying
PcpCache.

This restores the behavior of GetLoadSet in this case
to what it was prior to the introduction of the TF_VERIFY.
This behavior is confusing, though. A good follow-up
might be to unload prims whenever we deactivate them,
but that's a larger change.

(Internal change: 1667228)