Pcp: Dependency system redesign. Replace Spec-based tracking.
As Pcp indexes scene description it discovers composition arcs
representing dependencies ("deps") on other scene description.
Pcp must track deps to manage data lifetime, to process later
changes to scene data, to invalidate caches, and to support
features that must reason about the dependency graph structure,
such as namespace editing.

Each dependency points to a "site" representing all the SdfSpecs
("specs") that reside at a path in a layer stack.
Historically our old composition system tracked deps in terms
of spec stacks, not sites.  We retained that in Pcp, while
suspecting that tracking specs is a mismatch for the task,
because of the principle that composition arcs connect sites,
not specs, and the specs can always be derived from sites.

Further experience confirmed there were some issues.
The usd project found that spec tracking was too expensive
in memory terms, so it turns them off entirely, but at the
cost of not supporting all dependency features correctly.
The other issue is complexity: there are cases that
spec deps are insufficient to handle, so Pcp_Dependencies
was actually a hybrid that also tracked sites, and would
use one representation or the other in various cases.
This made the code tricky to follow, and required additional
concepts to emerge in the API like "fallback ancestor".

The new system tracks the sites used by every prim index,
and stores a reverse mapping from sites back to index paths.

details

* PcpDependencyType is now a full-fledged categorization
  of dependency types.  PcpClassifyNodeDependency()
  can compute the dep type of a given node.

* Rather than maintaining a self-contained external bimap,
  Pcp_Dependencies now takes advantage of lock-step update
  with PcpPrimIndex entries in PcpCache so it can store
  only the inverse map and use indexes for the forward map.
  This cuts the runtime memory usage in half.

* The above change means we can simplify the update process
  and no longer need an async "flush" pass.

* PcpPrimIndex no longer builds a side list of deps as it goes.
  Instead, we rely on the node graph to fully represent the
  actual dependency structure.

  To support this, we now introduce variant nodes for
  variant selections even when the variant specs at the
  target do not yet exist.  This lets us track the dependency
  purely as a node, and is consistent with other kinds of arcs.

  To handle "spooky dependencies" we now propagate implied
  relocate nodes up to the root -- formally representing
  what was previously implicit structure, and allowing
  "spooky" cases to be handled by the normal code paths.

  This obviates the "shouldContributeDependencies" node flag.
  "fallbackAncestor" is also no longer a necessary API concept,
  because we always have ancestral site deps to use.

* PcpLayerStackSite now holds a ref rather than weakptr to
  its layer stack.  This is consistent with its purpose
  in representing a dependency on a layer stack (not a
  weak reference!); allows us to avoid having a duplicate
  side-structure for tracking the actual references
  already described by the sites; and shrinks the size
  of a pcp graph node (48 B -> 40 B).  (It doesn't make
  any difference to the # of atomic ref count ops we
  do since weakptrs refcount their remnant; now we
  refcount the layerstack instead.)

results

* less code, fixes correctness bugs
* less runtime memory -- ~5-7x less Pcp dependency memory;
  one production set usd goes from 933MB to 883MB (total)
* depenency tracking overhead now scales with dependency
  arc complexity of the scene, not the prim count

bugs: 135974, 137333, 96529

(Internal change: 1667509)