Workaround for Python include ordering issue
Move Python-related includes to the top and also
switch the ordering. This is a minimal fix for build
issues on OSX due to the requirement that Python
headers always be included first. Ideally, we would
find some way to remove these headers from
vt/value.h entirely so that we don't impose a
requirement that vt/value.h be included first
as well.

Fixes #89

(Internal change: 1668848)