Add common delegate registration method to Hydra.
With the Hydra refactor, we are adding the ability for Hydra to support backend delegates supplied by plugins.

While TfType and Plug libraries provide discovery and loading mechanism, this checkin builds on their functionality in order to provide Hydra registration support.

This functionality has been introduced in a new library, called hf, which is a base library for the hydra eco-system (1 level above arch).

Changes:
- Added the Hf Library to the imaging/lib build scripts
- Added PerfLog.h, eventually the version from Hd should move here.
- Added PluginDelegateDesc.h which contains a structure that describes registed delegates to the application.
- Plugin Delegate Registry manages all registered delegates of a specified type.
- Plugin Delegate Entry is a private class that manages a single registered delegate in the registry.
- Plugin Delegate Base provides a simple base class, that is registered with TfType, for delegates that are provided to hydra through a plugin.

Notes:
This is all dead code at this time, other code has been written to exercise and test these additions, but will be checked in later.

(Internal change: 1674188)
(Internal change: 1674205)