always use the USD file being imported and primPaths based on that file when creating assemblies
This allows opinions present in the layer that contains the prim with the
reference to have an effect. By re-writing the reference to point directly
to asset file/prim paths, we may be losing opinions present in the
referencing layer (e.g. animation applied to the asset).

Also, this change addresses an issue with nested assemblies where the
namespace of members of the assembly would look something like
"NS_dagAsset1" rather than being based off of the assembly name. We were
creating and then immediately renaming nested assembly nodes using an
MDagModifer, but that rename operation was apparently too late and had no
effect on the nested assembly namespace that Maya automatically creates.
As a result, we rename the assembly node immediately after adding the
createNode operation and that seems to result in the correct node and
namespace names.

(Internal change: 1674255)