enable readAnimData when importing USD via assembly representation activations
This change enables the readAnimData import argument when activating
the Expanded and Full representations of an assembly so that animation gets
imported.

We also expand Maya's frame range to include a stage's frame range rather
than explicitly setting it to match when importing the stage. This makes it
so that multiple imports will result in the frame range containing the
union of all imported stages rather than just the range of the most recent
import.

(Internal change: 1674604)