Add Gal delegate registration to Hydra.
With the Hydra refactor, we are adding the ability for Hydra to support backend delegates supplied by plugins.

This checkin builds on proir work to add Gal as a seperate library and adds the ability to register Gal delegates.

The Gal delegate and this library are just place holders at the moment, with the primary motivation for implementation was testing the plugin delegate registration API supported multiple delegates and also provide the Hd Context creation with a Gal registry and Delegate types without an actual delegate being used.

Changes:
- Added the Gal Library to the imaging/lib build scripts
- Added GalDelegateRegistry as a specialization of HfPluginDelegateRegistry.
- Added GalDelegate baseclass which will be used to buildup the Gal API.

Notes:
This is all dead code at this time, other code has been written to exercise and test these additions, but will be checked in later.

(Internal change: 1674720)
(Internal change: 1674913)