connect the time attribute on assembly nodes to their proxies when activating hierarchical representations
This change also converts the "time" attribute on assembly and proxy nodes
to be time-typed rather than numerically typed.

With this change, all that's necessary to support animation with USD proxies
underneath assemblies is to connect Maya's global time (time1.outTime) to
the time attribute on assembly nodes. We do NOT do this by default as there
can be a significant performance penalty when the number of assembly nodes
grows large.

(Internal change: 1676141)