Katana can now import usd shading that uses a specializes hierarchy into a corresponding katana material hierarchy.
The way this works is that as the katana op walks down usd
namespace, it checks for child material prims that use
"specializes" arcs.  When it finds those, it inserts them
into a MaterialHierarchy, which is basically just a
base<->derived map.  It uses that map to re-splice the
materials into the right place for katana attribute inheritance
to do the right thing for base -> derived materials.

The other detail is that when buliding katana attributes,
we skip any whose resolved value originates beyond a
specializes arc.  This uses a set of helpers that
analyze the UsdResolveInfo and/or the opinion stacks.

This changelist is primarily meant as an existence proof that the
usd api has what we need to start making use of specializes arcs,
we expect to want to continue to refactor this and improve API's.

It's an open question whether this leaves too much of the
work on the client to interpret specializes, and if so what
API extensions Usd should provide to facilitate this.  But this
changelist at least demonstrates that we have the fundamentals
available, and hopefully provides a starting point for discussion
of more ergonomic API's.

details:
- analyze root material prims for specializes arcs and assemble
  the hierarchy up-front
- pass that hierarchy down the op chain in PrivateData
- use hierarchy info to splice child materials as children
  of their base materials
- skip emitting shading attributes that inherit from the base

(Internal change: 1676464)
(Internal change: 1676704)
(Internal change: 1676708)
(Internal change: 1676740)