Add ability to disable standard plugin loading for tests.
While trying to write a test for the new plugin discovery mechanisms in Hydra, I ran into an issue where real plugins on the search path interfered with the test.

In order to address this issue, this change adds a environment variable to plug that can disable loading plug-ins on this search path.

(Internal change: 1677103)