Add Context Creation to Hydra Engine.
This is part of the Hydra refactor.

This changes add's a Context Creation and support API to HdEngine.  In order to support advanced context creation this requires API's to create render indexes and discover back-end delegates.

The plugin management API provides functions for quering delegates and overriding the default delegate.

An API is provided for Adding/Inc Ref Count and Deleting a Render Index.  This is when an application was to use the fully specified context.

The Context Management API provides various ways to create contexts (default is the simplest, shared will clone an existing context or a fully specified context).

Additional Changes:
- Added library dependancies on hf and gal to build scripts.
- Added HdContext / Render Index Cache
- Added new HdEngine test to test the new API.
- Removed Render Delegate test as it is superseded by new test.

Notes:
This is all dead code at this time, other code has been written to exercise and test these additions, but will be checked in later.

(Internal change: 1677128)