Use ArResolver to get format extension for layer
Use ArResolver to get layer format extension instead
of TfStringGetSuffix in SdfLayer::_CreateNew, since
Ar gives implementations a chance to customize
behavior. This is consistent with the rest of
SdfLayer (except for CreateAnonymous -- see
comments for more details).

This was accomplished by refactoring the affected
code to use a pre-existing helper function.

(Internal change: 1677713)