Remove surfaceShaderId from HdRprim constructor
This is part of the Hydra refactor.

This change is a clean-up task in prep for some of the refactor work.

CL 1597959 changed the behavior of the way the surface shader id got passed to Hydra.  Prior to that change the surface shader id was provided in the Rprim constructor, however after that change the surface shader id passed in the constructor is ignored and Hydra always uses the id fetched from the delegate (or the fallback shader).

So this change removes the unused parameter from the HdRprim constructor.  However, as a near-future change will change the RenderIndex InsertRprim function, it has been left there as an unused parameter to maintain api compatability for now and avoid needing to port all the hydra client's twice.

Changes:
- Removed  surfaceShaderId from HdRprim, HdMesh, HdBasisCurves and HdPoints constructors.
- Updated RenderIndex to remove surfaceShaderId  from InsertRprim (but leave the argument placeholder)

(Internal change: 1678049)