Add support for getting textures from a draw target.
Previous changes for draw targets did not provide a any connection mechanisms between surface shaders and draw targets.  Surface shaders are unaware that a draw target is any different to a standard texture.

In order to enable this, this change exploits some of the existing texture mechanism:
When adding a texture to the scene, the texture sync mechanism uses the scene delegate to convert the texture prim's SdfPath into Id, for example by hashing that Id.  The texture then uses this path to insert a HdTextureResource derived class into the resource registry.  This HdTextureResource class provides an interface to obtain the GL resources necessary for the texture.

When a surface shader is bound to a texture, the shader params contain the SdfPath of the texture prim to bind to.  The surface shader also uses the scene delegate to turn this path into an ID and uses that to obtain the HdTextureResource from the registry.

The problem with draw target's is that a single prim represents multiple textures.  So to work around this problem, each attachment is addressed as a Property on the prim.

When the draw target's attachment set changes and a new draw target object is created, HdxDrawTarget creates a new Hdx_DrawTargetTextureResource and inserts it into the resource registry using the same path to id scene delegate function.  Each Hdx_DrawTargetTextureResource  stores a reference to the draw target attachment it represents.  Hdx_DrawTargetTextureResource implementation of the  HdTextureResource interface, uses the attachment information to return the appropriate GL resources from the draw target.

Another missing piece was the sampler state to use for drawing from those texture.  For now, the existing data structures are being used to piggyback that state.

Please Note: The Draw Target prim and support classes are present to aid transition of some internal tools to hydra and it not considered part of the core hydra concepts.

(Internal change: 1690631)