Pcp: Fix erroneous extra dependencies being registered (and not cleaned up later) in cases when authoring the first inert spec in a layer stack already used by a prim.
Parallel indexing wants to avoid checking the cache as
much as possible, to minimize synchronization.  Because
this kind of edit can invalidate a parent prim index but not
its children, parallel indexing needs to keep checking the cache
until it walks off the pathtable tree.

The reason this edit invalidates a parent but not children
is that it may cause a culled node to be unculled.
The corresponding nodes in the children remain unculled,
hence they are not invalidated.  See
PcpChanges::DidChangeSpecs() for more details.

This approach fixes the case in this bug while retaining
the performance of bringing up a scene for the first time
(i.e. when the cache is entry).  It also retains performance
when making large (aka significant) scene edits, since
they erase entire subtrees from the pathtable -- see
PcpCache::_RemovePrimAndPropertyCaches().

(Internal change: 1691353)