[usdMaya] allow tagging attributes for translation from double precision in Maya to single precision in USD
There is not always a direct mapping between Maya-native types and USD/Sdf
types, and often it's desirable to intentionally use a single precision
type when the extra precision is not needed to reduce size, bandwidth, etc.
For example, there is no native Maya attribute type to represent an array
of float triples. Now with this change, to get an attribute with a
VtVec3fArray type in USD, you can create a 'vectorArray' data-typed
attribute in Maya (which stores MVectors, which contain doubles) and set
'translateMayaDoubleToUsdSinglePrecision' to true so that the data is cast
to single-precision on export. It will be up-cast back to double on
re-import (when import is implemented).
By default, this property is false so that double precision data is
preserved in the translation back and forth between Maya and USD.

This change includes a new column in the Maya attribute tagger widget to
set the 'translateMayaDoubleToUsdSinglePrecision' property.

There were also some problems in our handling of a few different types of
Maya attributes, so this change also includes a clean-up of the
type-handling logic.

(Internal change: 1691664)