Fix for rpaths on OSX
An earlier fix caused the build to start setting
a colon-delimited list of $ORIGIN-based rpaths for
Python modules on OSX. This is broken for two
reasons: the OSX loader does not use $ORIGIN,
and it doesn't parse colon-delimited paths.

This fix makes it so we only set these rpaths
on Linux. This gets OSX builds working as they had
before while we continue working on this issue
as part of porting.

(Internal change: 1692521)