Delegate Sprim allocation to Render Delegate
This change is part of the Hydra refactor.

Following up on the Rprim changes, this change applies the same two transforms to Sprim's:
- Change renderindex storage from shared_ptr to raw pointer.
- Delegate Allocation/Dealaction to the Render Delegate.

In addition, this change modifies the Render Index to store Sprim's by type. User's of the Sprim API are usually interested in two things: Getting an sprim of a specific type and getting all sprims of a specific type.

With the old API, users usually had to query the returned Sprim to make sure it was the type they were expecting (usually via a dynamic cast). This was worse in the getting all prims cases, as more prims were visited to build the list and each item in the list needed individual checking. Making these get all prims very expensive.

Changes:
- Sprims are now stored in an unordered map keyed by typeId.
- All Sprim functions on render index now take a typeId as the first parameter.
- Sprim has a new virtual method, GetInitialDirtyBitsMask that returns the dirty bits desired for the first sync. The mirrors Rpim.
- Many users of Sprims changed from using the generic Sprim shared pointer to the typed hdx raw pointer for the prim type

(Internal change: 1697695)