Relative plugin paths for build relocatability
The build had been using absolute paths when
specifying built-in search paths. This made it very
difficult to relocate the build, since Plug would
always be looking at those paths no matter where
the build was moved to.

To fix this, the build now uses relative paths instead.
Plug has been updated to support relative paths for all
plugin search paths. Such paths are assumed to be relative
to the location of the Plug shared library.

This should make the build fully relocatable, so long
as the build products are moved as one monolithic
entity.

(Note that there are rpath issues on OSX and other
issues on Windows that will need to be fixed to make
this true on those platforms)

(Internal change: 1697700)